Moved syntactic parse tree types to central types.ts
