+++ frontend/src/app/services/parse.service.ts
@@ -2,8 +2,8 @@ import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
 import { Subject, switchMap, catchError, of, merge, map } from 'rxjs';
 import { ParseInput } from '@/annotate/annotation-input/annotation-input.component';
-import { ParseResponseData } from '@/annotate/annotation-parse-results/types';
 import { ProblemService } from './problem.service';
+import { ParseResponseData } from '@/types';
 
 
 export type ParseResponse = {
