+++ frontend/src/app/types.ts
@@ -132,5 +132,36 @@ export interface Dimensions {
     height: number;
 }
 
+//
+// Syntactic parse tree types
+//
 
-export type CCGTerm = string[];
+export type LeafNode = {
+    // Fixed order: rule, token, lemma, POS tag, NER tag, category.
+    node: [string, string, string, string, string, string];
+};
+
+export type UnaryNode = {
+    node: string;
+    children: [CCGNode];
+};
+
+export type BinaryNode = {
+    node: string;
+    children: [CCGNode, CCGNode];
+};
+
+export type CCGNode = LeafNode | UnaryNode | BinaryNode;
+
+export type ParseTreeType = 'ccg_tree' | 'ccg_term' | 'corr_term' | 'llf';
+
+export interface CCGParse {
+    sentence: string;
+    ccg_trees: Record<ParseTreeType, CCGNode>;
+};
+
+export type ParseResponseData = {
+    ccg_parses: CCGParse[];
+    proofs: unknown[];
+};
+
